Fix all directories
