+++ src/CourseSearch/components/SearchByCoreCurriculumAndFlags.jsx
