+++ src/CourseSearch/components/SearchByDepartmentAndLevel.jsx
