+++ src/CourseSearch/styles/SearchByCoreCurriculumAndFlags.css
