+++ src/DegreeDuoCalendar/SideBar/components/CourseDay.jsx
