+++ src/GuestPopUp.jsx
@@ -1,7 +1,7 @@
 import { useNavigate } from 'react-router-dom';
 import './styles/GuestPopUp.css';
-import BackComponent from './Components/BackComponent';
-import Title from './Components/Title';
+import BackComponent from './Universal Components/BackComponent';
+import Title from './Universal Components/Title';
 
 function GuestProfile() {
   const navigate = useNavigate();
