+++ src/Register.jsx
@@ -1,6 +1,6 @@
 import { useNavigate } from 'react-router-dom';
-import Title from './Components/Title';
-import BackComponent from './Components/BackComponent';
+import Title from './Universal Components/Title';
+import BackComponent from './Universal Components/BackComponent';
 import './styles/Register.css';
 
 const majors = [
