--- src/styles/SearchByCoreCurriculumAndFlags.css
@@ -1,77 +1,0 @@
-.search-row {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-  
-  .search-bar {
-    margin-bottom: 20px;
-  }
-  
-  .dropdown {
-    width: 300px;
-    height: 40px;
-    font-size: 16px;
-    padding: 5px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-  }
-  
-  .checkbox-group {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    gap: 10px;
-    margin-bottom: 20px;
-  }
-  
-  .checkbox-wrapper {
-    display: flex;
-    align-items: center;
-    cursor: pointer;
-    border: 2px solid #FFA500; /* Orange border */
-    border-radius: 5px;
-    padding: 5px;
-    background-color: #fff;
-  }
-  
-  .checkbox-wrapper.selected {
-    background-color: #FFA500; /* Orange background when selected */
-  }
-  
-  .checkbox {
-    width: 20px;
-    height: 20px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: #fff;
-    border: 1px solid #ccc;
-    margin-right: 5px;
-    border-radius: 3px;
-  }
-  
-  .checkmark {
-    font-size: 16px;
-    color: #000; /* Black check mark */
-  }
-  
-  .checkbox-label {
-    font-size: 16px;
-    color: #000;
-  }
-  
-  .search-button {
-    background-color: #bf5701;
-    color: #fff;
-    font-size: 16px;
-    padding: 10px 20px;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-  }
-  
-  .search-button:hover {
-    background-color: #a84b01;
-  }
-
